\name{achievement}
\alias{achievement}
\docType{data}
\title{Achievement tests for children}
\description{Simulated dataset containing the score on four achievement tests for 50 childrens attending the last year of primary school. The four tests are repeated five times at regular time intervals.}
\usage{data(achievement)}
\format{
  A data.frame with a total of 250 observations on the following 6 variables:
  \describe{
    \item{\code{id}}{Child identifier.}
    \item{\code{time}}{Assessment time (1 to 5).}
    \item{\code{speaking}}{Score on speaking test (1=minimum, 5=maximum).}
    \item{\code{reading}}{Score on reading test (1=minimum, 5=maximum).}
    \item{\code{writing}}{Score on writing test (1=minimum, 5=maximum).}
    \item{\code{math}}{Score on math test (1=minimum, 5=maximum).}
    }
  }
%\note{}
%\references{}
\seealso{\link{achievement}}