\name{agrisus}
\alias{agrisus}
\docType{data}
\title{EU agricultural sustainability data}
\description{Data on several indicators covering the economic, social and environmental dimensions of agricultural sustainability for 25 EU countries in the period 2004-2018.}
\usage{data(agrisus)}
\format{
  A data.frame with a total of 375 observations on the following 15 variables:
  \describe{
    \item{\code{}}{.}
    \item{\code{}}{.}
    \item{\code{}}{.}
    \item{\code{}}{.}
    \item{\code{}}{.}
    \item{\code{}}{.}
    \item{\code{}}{.}
    \item{\code{}}{.}
    \item{\code{}}{.}
    \item{\code{}}{.}
  }
}
