\name{agrisus}
\alias{agrisus}
\docType{data}
\title{EU agricultural sustainability data}
\description{Data on several indicators covering the economic, social and environmental dimensions of agricultural sustainability for 25 EU countries in the period 2004-2018.}
\usage{data(agrisus)}
\format{
  A data.frame with a total of 375 observations on the following 15 variables:
  \describe{
    \item{\code{Country}}{Country name.}
    \item{\code{Country_code}}{Country code.}
    \item{\code{Year}}{Year.}
    \item{\code{TFP\_2005\_CAP}}{Total Factor Productivity (TFP) index of agriculture (2005=100). Source: Eurostat.}
    \item{\code{NetCapital\_GVA}}{Net capital stocks in agriculture (2015 US dollars) to gross value added of agriculture (2015 US dollars). Source: FAO.}
    \item{\code{Manager\_ratio}}{Ratio young/elderly for farm managers (number of managers with less than 35 years by 100 managers with 55 years and over). Source: Eurostat.}
    \item{\code{FactorIncome\_paid\_2010}}{Real income of agricultural factors per paid annual work unit (index 2010=100). Source: Eurostat.}
    \item{\code{EntrIncome\_unpaid\_2010}}{Net entrepreneurial income of agriculture per unpaid annual work unit (index 2010=100). Source: Eurostat.}
    \item{\code{Income\_rur}}{Median equivalised net income in rural areas (power purchasing standard). Source: Eurostat.}
    \item{\code{Unempl\_rur}}{At-risk-of-poverty rate in rural areas (\%). Source: Eurostat.}
    \item{\code{Poverty\_rur}}{Unemployment rate in rural areas (\%). Source: Eurostat.}
    \item{\code{RenewProd}}{Production of renewable energy from agriculture (share of total production of renewable energy, \%). Source: Eurostat.}
    \item{\code{Organic\_p}}{Area under organic cultivation (share of utilized agricultural area, \%). Source: Eurostat.}
    \item{\code{GHG\_UAA}}{Greenhouse gas emissions due to agriculture (million CO2 equivalent grams per hectare of utilized agricultural area). Source: FAO.}
    \item{\code{GNB\_UAA}}{Gross nitrogen balance (tonnes of nutrient per hectare of utilized agricultural area). Source: OECD.}
  }
}
